+++ backend/services/question/question.js
@@ -2,6 +2,7 @@ const {GDBQuestionModel} = require("../../models/ClientFunctionalities/question"
 const {createQuestionHelper, updateQuestionHelper, findQuestionById} = require("./questionHelper");
 const {MDBDynamicFormModel} = require("../../models/dynamicForm");
 const {GDBClientModel, GDBOrganizationModel} = require("../../models");
+const {GraphDB} = require("../../utils/graphdb");
 
 
 const createQuestion = async (req, res, next) => {
@@ -28,26 +29,37 @@ const updateQuestion = async (req, res, next) => {
 
   try {
     const forms = await MDBDynamicFormModel.find({formStructure: {$elemMatch: {fields: {$elemMatch: {id: id, type: 'question'}}}}})
-    const clients = (await GDBClientModel.find({}, {populates: ['questionOccurrences.occurrenceOf']})).filter((client) => {
-      if(client.questionOccurrences){
-        for (let occurrence of client.questionOccurrences) {
-          if (occurrence.occurrenceOf._id === id)
-            return true
-        }
-      }
-      return false
-    })
-    const organizations = (await GDBOrganizationModel.find({}, {populates: ['questionOccurrences.occurrenceOf']})).filter((organization) => {
-      if(organization.questionOccurrences){
-        for (let occurrence of organization.questionOccurrences) {
-          if (occurrence.occurrenceOf._id === id)
-            return true
-        }
-      }
-      return false
-    })
+    // const clients = (await GDBClientModel.find({}, {populates: ['questionOccurrences.occurrenceOf']})).filter((client) => {
+    //   if(client.questionOccurrences){
+    //     for (let occurrence of client.questionOccurrences) {
+    //       if (occurrence.occurrenceOf._id === id)
+    //         return true
+    //     }
+    //   }
+    //   return false
+    // })
+    // const organizations = (await GDBOrganizationModel.find({}, {populates: ['questionOccurrences.occurrenceOf']})).filter((organization) => {
+    //   if(organization.questionOccurrences){
+    //     for (let occurrence of organization.questionOccurrences) {
+    //       if (occurrence.occurrenceOf._id === id)
+    //         return true
+    //     }
+    //   }
+    //   return false
+    // })
+    const query = `
+    PREFIX : <http://snmi#>
+    select * where { 
+	      ?s ?p :question_${id}.
+        ?s a :QuestionOccurrence.
+    } limit 1`
+
+    let isUsed = false
+    await GraphDB.sendSelectQuery(query, false, () => {
+      isUsed = true
+    });
     const question = await findQuestionById(id);
-    if(forms.length !== 0 || clients.length !== 0 || organizations.length !== 0 || question.isPredefined)
+    if(forms.length !== 0 || isUsed || question.isPredefined)
       res.status(400).json({success: false, message: 'This question cannot be updated'})
     await updateQuestionHelper(id, updateData);
     return res.status(202).json({success: true, message: 'Successfully update question.'});
@@ -62,25 +74,36 @@ const fetchQuestion = async (req, res, next) => {
     const id = req.params.id;
     const question = await findQuestionById(id);
     const forms = await MDBDynamicFormModel.find({formStructure: {$elemMatch: {fields: {$elemMatch: {id: id, type: 'question'}}}}})
-    const clients = (await GDBClientModel.find({}, {populates: ['questionOccurrences.occurrenceOf']})).filter((client) => {
-      if(client.questionOccurrences){
-        for (let occurrence of client.questionOccurrences) {
-          if (occurrence.occurrenceOf._id === id)
-            return true
-        }
-      }
-      return false
-    })
-    const organizations = (await GDBOrganizationModel.find({}, {populates: ['questionOccurrences.occurrenceOf']})).filter((organization) => {
-      if(organization.questionOccurrences){
-        for (let occurrence of organization.questionOccurrences) {
-          if (occurrence.occurrenceOf._id === id)
-            return true
-        }
-      }
-      return false
-    })
-    return res.status(200).json({question, success: true, locked: forms.length !== 0 || clients.length !== 0 || organizations.length !== 0 || question.isPredefined});
+    // const clients = (await GDBClientModel.find({}, {populates: ['questionOccurrences.occurrenceOf']})).filter((client) => {
+    //   if(client.questionOccurrences){
+    //     for (let occurrence of client.questionOccurrences) {
+    //       if (occurrence.occurrenceOf._id === id)
+    //         return true
+    //     }
+    //   }
+    //   return false
+    // })
+    // const organizations = (await GDBOrganizationModel.find({}, {populates: ['questionOccurrences.occurrenceOf']})).filter((organization) => {
+    //   if(organization.questionOccurrences){
+    //     for (let occurrence of organization.questionOccurrences) {
+    //       if (occurrence.occurrenceOf._id === id)
+    //         return true
+    //     }
+    //   }
+    //   return false
+    // })
+    const query = `
+    PREFIX : <http://snmi#>
+    select * where { 
+	      ?s ?p :question_${id}.
+        ?s a :QuestionOccurrence.
+    } limit 1`
+
+    let isUsed = false
+    await GraphDB.sendSelectQuery(query, false, () => {
+      isUsed = true
+    });
+    return res.status(200).json({question, success: true, locked: forms.length !== 0 || isUsed || question.isPredefined});
   } catch (e) {
     next(e)
   }
@@ -105,6 +128,42 @@ const fetchQuestions = async (req, res, next) => {
 const deleteQuestion = async (req, res, next) => {
   try {
     const id = req.params.id;
+    const question = await findQuestionById(id);
+    // find out the forms linked to the question
+    const forms = await MDBDynamicFormModel.find({formStructure: {$elemMatch: {fields: {$elemMatch: {id: id, type: 'question'}}}}})
+    // const clients = (await GDBClientModel.find({}, {populates: ['questionOccurrences.occurrenceOf']})).filter((client) => {
+    //   if(client.questionOccurrences){
+    //     for (let occurrence of client.questionOccurrences) {
+    //       if (occurrence.occurrenceOf._id === id)
+    //         return true
+    //     }
+    //   }
+    //   return false
+    // })
+    // // find out the organizations linked to the question
+    // const organizations = (await GDBOrganizationModel.find({}, {populates: ['questionOccurrences.occurrenceOf']})).filter((organization) => {
+    //   if(organization.questionOccurrences){
+    //     for (let occurrence of organization.questionOccurrences) {
+    //       if (occurrence.occurrenceOf._id === id)
+    //         return true
+    //     }
+    //   }
+    //   return false
+    // })
+    const query = `
+    PREFIX : <http://snmi#>
+    select * where { 
+	      ?s ?p :question_${id}.
+        ?s a :QuestionOccurrence.
+    } limit 1`
+
+    let isUsed = false
+    await GraphDB.sendSelectQuery(query, false, () => {
+      isUsed = true
+    });
+
+    if(question.isPredefined || forms.length !== 0 || isUsed)
+      return res.status(400).json({success: false, message: 'The question is not deletable'});
     const doc = await GDBQuestionModel.findByIdAndDelete(id);
     return res.status(200).json({success: true});
   } catch (e) {
